Reduce dependence on %ENV for internal code.
Perl 5.17.3+ changes behaviour of `%ENV` so values are also stringified.

This change aims to solve that ( gh #870 ) by replacing %ENV based code with
a passed around local hashref, that is passed as the last parameter to `new_for_request()`
instead of relying on the data being passed around via `%ENV`.

Conflicts:

	lib/Dancer/Request.pm
	lib/Dancer/Test.pm